# core/bloombits: faster generator