This PR optimizes the bloombit generator a bit. It unrolls a loop and avoids branching on every potential setting of a bit. 
EDIT: It also performs drastically better when blooms are less saturated

```
name                old time/op    new time/op    delta
Generator/empty-6     7.72ms ± 1%    0.91ms ± 3%  -88.27%  (p=0.000 n=9+8)
Generator/random-6    28.9ms ± 1%    15.2ms ± 1%  -47.44%  (p=0.000 n=9+9)

```
The existing testcase was pretty good. I tested with some intentional (and unintentional) errors, and each one was immediately detected.